+++ lighthouse-plugin-publisher-ads/audits/critical-blocking-requests.js
@@ -12,14 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-const AdRequestCriticalPath = require('./ad-request-critical-path');
 const i18n = require('lighthouse/lighthouse-core/lib/i18n/i18n');
-const NetworkRecords = require('lighthouse/lighthouse-core/computed/network-records');
 const {auditNotApplicable} = require('../messages/common-strings');
 const {Audit} = require('lighthouse');
-const {getAbbreviatedUrl, trimUrl} = require('../utils/resource-classification');
-const {getAdCriticalGraph} = require('../utils/graph');
-const {getTimingsByRecord} = require('../utils/network-timing');
+const {computeAdRequestWaterfall} = require('../utils/graph');
 
 /** @typedef {LH.Artifacts.NetworkRequest} NetworkRequest */
 /** @typedef {LH.Gatherer.Simulation.NodeTiming} NodeTiming */
@@ -38,16 +34,6 @@ const UIStrings = {
 
 const str_ = i18n.createMessageInstanceIdFn(__filename, UIStrings);
 
-/**
- * @typedef {Object} SimpleRequest
- * @property {string} url
- * @property {string} abbreviatedUrl
- * @property {string} type
- * @property {number} startTime
- * @property {number} endTime
- * @property {number} duration
- * @property {?number?} selfTime
- */
 
 /**
  * Table headings for audits details sections.
@@ -73,43 +59,6 @@ const HEADINGS = [
   },
 ];
 
-/**
- * @param {SimpleRequest[]} requests A pre-sorted list of requests by start
- *   time.
- * @return {SimpleRequest[]}
- */
-function findBlockingRequests(requests) {
-  let bottlneckRequest = requests[0];
-  bottlneckRequest.selfTime = bottlneckRequest.duration;
-
-  let scanEnd = bottlneckRequest.startTime;
-
-  const results = [];
-  for (const current of requests) {
-    if (current.endTime < scanEnd || current == bottlneckRequest) {
-      // Overlaps with previous requests, skip to avoid double counting.
-      continue;
-    }
-    const left = Math.max(scanEnd, current.startTime);
-    const right = Math.min(bottlneckRequest.endTime, current.endTime);
-    if (left < right) {
-      // @ts-ignore selfTime is initialized elsewhere, so it won't be undefined.
-      bottlneckRequest.selfTime -= (right - left);
-    }
-    scanEnd = Math.max(scanEnd, right);
-    if (current.endTime > bottlneckRequest.endTime) {
-      // The next request is a potential bottleneck.
-      if (bottlneckRequest.selfTime > 100) {
-        bottlneckRequest.initiatorUrl = bottlneckRequest.initiatorRequest ?
-          bottlneckRequest.initiatorRequest.url : '',
-        results.push(bottlneckRequest);
-      }
-      current.selfTime = current.endTime - left;
-      bottlneckRequest = current;
-    }
-  }
-  return results;
-}
 
 /**
  * Audit to check the length of the critical path to load ads.
@@ -136,25 +85,30 @@ class CriticalBlockingRequests extends Audit {
    * @return {Promise<LH.Audit.Product>}
    */
   static async audit(artifacts, context) {
+    const trace = artifacts.traces[Audit.DEFAULT_PASS];
+    const devtoolsLog = artifacts.devtoolsLogs[Audit.DEFAULT_PASS];
+
     const waterfall =
-      await AdRequestCriticalPath.computeResults(artifacts, context);
+      await computeAdRequestWaterfall(trace, devtoolsLog, context);
     if (!waterfall.length) {
       return auditNotApplicable.NoAdRelatedReq;
     }
-    const criticalRequests = findBlockingRequests(waterfall);
-    // Only show the top critical requests for the sake of brevity.
-    const topCriticalRequests = criticalRequests
+    const CRITICAL_SELF_TIME_MS = 200;
+    const criticalRequests = waterfall
+        .filter((a) => a.selfTime > CRITICAL_SELF_TIME_MS)
         .sort((a, b) => b.selfTime - a.selfTime)
+        // Only show the top critical requests for the sake of brevity.
         .slice(0, 5);
     const blockedTime =
-      topCriticalRequests.reduce((sum, r) => sum + r.selfTime, 0) / 1000;
+      // @ts-ignore param types not inferred.
+      criticalRequests.reduce((sum, r) => sum + r.selfTime, 0) / 1000;
     const failed = criticalRequests.length > 3 || blockedTime > 0.5;
     return {
-      numericValue: topCriticalRequests.length,
+      numericValue: criticalRequests.length,
       score: failed ? 0 : 1,
       displayValue: failed ? str_(UIStrings.displayValue, {blockedTime}) : '',
-      details: AdRequestCriticalPath.makeTableDetails(
-        HEADINGS, topCriticalRequests),
+      details:
+        CriticalBlockingRequests.makeTableDetails(HEADINGS, criticalRequests),
     };
   }
 }
